pyproject.toml
[tool.poetry]
name = "mysite"
version = "0.1.0"
description = ""
authors = ["yaseen1998 <Yasseenkouthe@ENG.HU.EDU.JO>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = ">=4.0,<4.1"
install = "^1.3.5"
wagtail = ">=3.0,<3.1"
Willow = "^1.4.1"
XlsxWriter = "^3.0.3"
django-debug-toolbar = "^3.4.0"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
